--- lib/presentation/schedule/navermap/view/navermap_content/navermap_searchbar.dart
@@ -1,8 +1,0 @@
-part of '../navermap_view.dart';
-
-final class _NaverMapSearchBar extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return const TextField();
-  }
-}
