+++ lib/presentation/schedule/searchplace/bloc/searchplace_bloc.dart
@@ -1,0 +1,2 @@
+part 'searchplace_event.dart';
+part 'searchplace_state.dart';
