+++ lib/presentation/schedule/searchplace/bloc/searchplace_event.dart
@@ -1,0 +1,1 @@
+part of 'searchplace_bloc.dart';
