+++ lib/presentation/schedule/searchplace/view/searchplace_content/searchplace_content.dart
@@ -1,23 +1,19 @@
-part of '../navermap_view.dart';
+part of '../searchplace_view.dart';
 
-final class _NaverMapContent extends StatelessWidget {
-  BuildContext childContext;
-
-  _NaverMapContent({required this.childContext});
-
+final class _SearchPlaceContent extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return SafeArea(
       bottom: false,
       child: Scaffold(
         appBar: AppBar(
-          title: _NaverMapSearchBar(),
+          title: _SearchPlaceSearchBar(),
           scrolledUnderElevation: 0,
           bottomOpacity: 0,
           elevation: 0,
           backgroundColor: Colors.transparent,
         ),
-        body: _NaverMapSearchResultView(),
+        body: _SearchPlaceListView(),
       ),
     );
   }
