+++ lib/presentation/schedule/searchplace/view/searchplace_content/searchplace_listview.dart
@@ -1,0 +1,18 @@
+part of '../searchplace_view.dart';
+
+final class _SearchPlaceListView extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return ListView(
+      shrinkWrap: true,
+      controller: ModalScrollController.of(context),
+      children: ListTile.divideTiles(
+        context: context,
+        tiles: List.generate(
+          100,
+          (index) => _SearchPlaceListItem(),
+        ),
+      ).toList(),
+    );
+  }
+}
