+++ lib/presentation/schedule/searchplace/view/searchplace_content/searchplace_searchbar.dart
@@ -1,0 +1,8 @@
+part of '../searchplace_view.dart';
+
+final class _SearchPlaceSearchBar extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return const TextField();
+  }
+}
