+++ lib/presentation/schedule/searchplace/view/view.dart
@@ -1,0 +1,1 @@
+export 'searchplace_view.dart';
